Feat/webpacking (#5)
* set up webpack.

This minimal webpack config allows es6 imports / exports (as far as I can tell)

It also spits out a compiled "scripts" bundle into the `dist` folder. This can be changed easily.

The command to get webpack running is currently: `npm run dev`

There is a possibility of using webpack-dev-server to enable hot reloading, but I'm not sure that makes sense since the app doesn't actually have an index.html as far as I can tell.

### resources ###
webpack reference came from  - webpack base config from: https://gist.github.com/learncodeacademy/25092d8f1daf5e4a6fd3

video I watched for getting some insight - video reference: https://www.youtube.com/watch?v=9kJVYpOqcVU

* Move JSS styles into a seperate file

The jss is now being spit out in the final dist/scripts.js file, but this is untested.

* Use already existing script.jsx for webpack entry + linting.

I removed the unecessary index.js file -- I originally used this thinking that I need to have some kind of "parent" level file that just handled imports and sent it off to webpack -- but on PR review, it makes more sense to use the scripts.jsx (already existing)

I also did some linting and watched some tests go green!
