+++ .gitignore
@@ -37,3 +37,7 @@ jspm_packages
 
 # Optional REPL history
 .node_repl_history
+
+
+# webpack output
+dist
