+++ src/script.jsx
@@ -1,3 +1,5 @@
+import styles from './styles';
+
 const script = (function script(w) {
   // Libraries
 
@@ -122,403 +124,7 @@ const script = (function script(w) {
 
   // TODO: This is ugly and I would really love to have these as seperate files
   //  or some way of not bloating the source file
-  const styles = {
-    MESSANGER: {
-      Message: {
-        operatorPicture: {
-          width: '48px',
-          height: '48px',
-          float: 'right',
-          marginTop: '4px',
-          boxSizing: 'border-box',
-          padding: '0 4px',
-        },
-        operatorPictureImage: {
-          width: '40px',
-          height: '40px',
-        },
-        operatorContent: {
-          margin: 0,
-          padding: '6px',
-          listStyle: 'none',
-          width: '160px',
-          marginTop: '4px',
-          marginBottom: '4px',
-          borderRadius: '10px 0 10px 10px',
-          background: '#e1e1e1',
-          float: 'right',
-          textAlign: 'right',
-        },
-        userContent: {
-          margin: 0,
-          listStyle: 'none',
-          width: '160px',
-          float: 'left',
-          textAlign: 'left',
-          background: '#0a6bef',
-          marginTop: '0',
-          marginLeft: '4px',
-          borderRadius: '10px 10px 10px 0',
-          padding: '6px',
-          color: 'white',
-        },
-      },
-      MessageList: {
-        messages: {
-          position: 'absolute',
-          top: '32px',
-          bottom: '48px',
-          left: 0,
-          right: 0,
-          paddingTop: '6px',
-          overflowY: 'scroll',
-          borderRight: '1px solid #ccc',
-          boxSizing: 'border-box',
-        },
-        messagesWrapper: {
-          position: 'relative',
-          height: '100%',
-        },
-        messagesList: {
-          margin: 0,
-          padding: 0,
-          listStyle: 'none',
-          fontSize: '14px',
-          fontFamily: 'sans-serif',
-          width: '100%',
-        },
-      },
-      Input: {
-        input: {
-          position: 'absolute',
-          bottom: 0,
-          left: 0,
-          right: 0,
-          width: '100%',
-          height: '48px',
-          boxSizing: 'border-box',
-          border: 0,
-          borderRight: '1px solid #ccc',
-          borderTop: '1px solid #ddd',
-          backgroundColor: '#fff',
-          padding: '8px',
-          color: '#222',
-          fontSize: '11px',
-          resize: 'none',
-          outline: 0,
-        },
-      },
-      Chat: {
-        innerWrapper: {
-          display: 'block',
-          width: '280px',
-          height: '360px',
-          boxShadow: '#dddddd 1px 1px 8px 0',
-          borderWidth: '0 1px 1px 0',
-          borderRadius: '3px 3px 0 0',
-          borderStyle: 'solid',
-          borderColor: '#cccccc',
-          borderBottom: 0,
-        },
-        outerWrapper: {
-          position: 'relative',
-          width: '80%',
-          margin: '0 auto',
-        },
-        box: {
-          position: 'absolute',
-          right: 0,
-          bottom: 0,
-          backgroundColor: '#ffffff',
-        },
-        header: {
-          display: 'block',
-          width: '100%',
-          height: '32px',
-          padding: '8px',
-          boxSizing: 'border-box',
-          borderRadius: '3px 3px 0 0',
-          boxShadow: '0 1px 1px 0 rgba(0,0,0,0.15)',
-          fontSize: '13px',
-          fontFamily: '\'Arial\', sans-serif',
-          color: 'white',
-          background: '#ef7f7f',
-        },
-        icon: {
-          border: 0,
-          float: 'right',
-          fontSize: '24px',
-          marginRight: '-8px',
-          marginTop: '-7px',
-          fontWeight: '800',
-          cursor: 'pointer',
-          color: 'rgba(255,255,255,0.5)',
-          background: 'rgba(0, 0, 0, 0)',
-        },
-      },
-    },
-    FLOAT: {
-      Message: {
-        operatorPicture: {
-          position: 'absolute',
-          left: '-48px',
-          width: '48px',
-          height: '48px',
-          float: 'right',
-          marginTop: '-16px',
-          boxSizing: 'border-box',
-          padding: '0 4px',
-        },
-        operatorPictureImage: {
-          width: '40px',
-          height: '40px',
-          borderRadius: '20px',
-        },
-        operatorContent: {
-          margin: 0,
-          listStyle: 'none',
-          width: '160px',
-          float: 'left',
-          textAlign: 'left',
-          background: '#ffffff',
-          marginTop: '4px',
-          marginLeft: '4px',
-          boxShadow: '#dddddd 1px 1px 8px 0',
-          borderRadius: '0 10px 10px 10px',
-          padding: '16px',
-        },
-        userContent: {
-          margin: 0,
-          padding: '16px',
-          listStyle: 'none',
-          width: '160px',
-          marginTop: '4px',
-          marginBottom: '4px',
-          boxShadow: '#dddddd 1px 1px 8px 0',
-          borderRadius: '10px 10px 0 10px',
-          background: '#0a6bef',
-          float: 'right',
-          textAlign: 'right',
-          color: 'white',
-        },
-      },
-      MessageList: {
-        messages: {
-          position: 'absolute',
-          bottom: '48px',
-          left: 0,
-          right: 0,
-          paddingTop: '6px',
-          paddingBottom: '6px',
-          boxSizing: 'border-box',
-        },
-        messagesWrapper: {
-          position: 'relative',
-          height: '100%',
-        },
-        messagesList: {
-          margin: 0,
-          padding: 0,
-          listStyle: 'none',
-          fontSize: '14px',
-          fontFamily: 'sans-serif',
-          width: '100%',
-        },
-      },
-      Input: {
-        input: {
-          position: 'absolute',
-          bottom: 0,
-          left: 0,
-          right: 0,
-          width: '100%',
-          height: '48px',
-          boxSizing: 'border-box',
-          boxShadow: '#dddddd 1px 1px 8px 0',
-          border: 0,
-          backgroundColor: '#fff',
-          borderRadius: '10px',
-          padding: '16px',
-          color: '#222',
-          fontSize: '13px',
-          resize: 'none',
-          outline: 0,
-        },
-      },
-      Chat: {
-        innerWrapper: {
-          display: 'block',
-          width: '280px',
-          height: '360px',
-        },
-        outerWrapper: {
-          position: 'relative',
-          width: '90%',
-          margin: '0 auto',
-        },
-        box: {
-          position: 'absolute',
-          right: 0,
-          bottom: '120px',
-          backgroundColor: 'transparent',
-        },
-        header: {
-          position: 'absolute',
-          bottom: '-48px',
-          right: '-48px',
-          display: 'block',
-          width: '42px',
-          height: '42px',
-          padding: '11px 5px 11px 6px',
-          boxSizing: 'border-box',
-          borderRadius: '21px',
-          boxShadow: '#dddddd 1px 1px 8px 0',
-          fontSize: '13px',
-          fontFamily: '\'Arial\', sans-serif',
-          color: 'white',
-          background: '#ef7f7f',
-        },
-        headerText: {
-          display: 'none',
-        },
-        icon: {
-          float: 'right',
-          fontSize: '24px',
-          marginTop: '-5px',
-          fontWeight: '800',
-          border: 0,
-          cursor: 'pointer',
-          color: 'rgba(255,255,255,0.5)',
-          background: 'rgba(0, 0, 0, 0)',
-        },
-      },
-    },
-    SIDEPANEL: {
-      Message: {
-        operatorPicture: {
-          width: '48px',
-          height: '48px',
-          float: 'right',
-          marginTop: '4px',
-          boxSizing: 'border-box',
-          padding: '0 4px',
-        },
-        operatorPictureImage: {
-          width: '40px',
-          height: '40px',
-        },
-        operatorContent: {
-          margin: 0,
-          padding: '6px',
-          listStyle: 'none',
-          width: '160px',
-          marginTop: '4px',
-          marginBottom: '4px',
-          borderRadius: '10px 0 10px 10px',
-          background: '#e1e1e1',
-          float: 'right',
-          textAlign: 'right',
-        },
-        userContent: {
-          margin: 0,
-          listStyle: 'none',
-          width: '160px',
-          float: 'left',
-          textAlign: 'left',
-          background: '#0a6bef',
-          marginTop: '0',
-          marginLeft: '4px',
-          borderRadius: '10px 10px 10px 0',
-          padding: '6px',
-          color: 'white',
-        },
-      },
-      MessageList: {
-        messages: {
-          position: 'absolute',
-          top: 0,
-          bottom: '48px',
-          left: 0,
-          right: 0,
-          paddingTop: '6px',
-          borderRight: '1px solid #ccc',
-          boxSizing: 'border-box',
-        },
-        messagesWrapper: {
-          position: 'relative',
-          height: '100%',
-        },
-        messagesList: {
-          margin: 0,
-          padding: 0,
-          listStyle: 'none',
-          fontSize: '14px',
-          fontFamily: 'sans-serif',
-          width: '100%',
-        },
-      },
-      Input: {
-        input: {
-          position: 'absolute',
-          bottom: 0,
-          left: 0,
-          right: 0,
-          width: '100%',
-          height: '48px',
-          boxSizing: 'border-box',
-          border: 0,
-          borderRight: '1px solid #ccc',
-          borderTop: '1px solid #ddd',
-          backgroundColor: '#fff',
-          padding: '8px',
-          color: '#222',
-          fontSize: '11px',
-          resize: 'none',
-          outline: 0,
-        },
-      },
-      Chat: {
-        innerWrapper: {
-          display: 'block',
-          width: '280px',
-          height: '100%',
-          boxShadow: '#dddddd 1px 1px 8px 0',
-          borderWidth: '0 1px 1px 0',
-          borderRadius: '3px 3px 0 0',
-          borderStyle: 'solid',
-          borderColor: '#cccccc',
-          borderBottom: 0,
-        },
-        outerWrapper: {
-          position: 'relative',
-          height: '100%',
-          width: '80%',
-          margin: '0 auto',
-        },
-        box: {
-          position: 'absolute',
-          top: 0,
-          bottom: 0,
-          right: 0,
-          backgroundColor: '#ffffff',
-        },
-        header: {
-          display: 'none',
-        },
-        icon: {
-          border: 0,
-          float: 'right',
-          fontSize: '24px',
-          marginRight: '-10px',
-          marginTop: '-7px',
-          fontWeight: '800',
-          cursor: 'pointer',
-          color: 'rgba(255,255,255,0.5)',
-          background: 'rgba(0, 0, 0, 0)',
-        },
-      },
-    },
-  };
+
   const style = function Style (component) {
     const state = store.getState();
